[BLINK-] Modify the parameter X of put[X]NormalizedKey(X, ...) methods in ComparatorUtil to primary type
Summary: fix T16820190

Reviewers: 玉兆

Reviewed By: 玉兆

Subscribers: 之信, P710480

Differential Revision: https://aone.alibaba-inc.com/code/D676879

+++ flink-core/src/main/java/org/apache/flink/api/common/typeutils/base/LongComparator.java
@@ -60,7 +60,7 @@ public final class LongComparator extends BasicTypeComparator<Long> {
 
 	@Override
 	public void putNormalizedKey(Long lValue, MemorySegment target, int offset, int numBytes) {
-		ComparatorUtil.putLongNormalizedKey(lValue, target, offset, numBytes);
+		ComparatorUtil.putLongNormalizedKey(lValue.longValue(), target, offset, numBytes);
 	}
 
 	@Override
